requirements.txt
streamlit==1.25.0
pandas==2.0.3
numpy==1.26.4
python-dotenv==1.0.1
protobuf==3.20.3